requirements.txt
aiogram
python-dotenv
PyPDF2
ebooklib
Pillow
